pipe: allocate and destroy pipe transport data with common pipe data
This updates the pipe to use contiguous data for the transport data
as well as the pipe protocol data.  It updates sockfd to use this, and
eliminates the need for the sockfd transport to do its own asynchronous
reaping, thereby hopefully closing a shutdown race.

The other transports will shortly get the same treatment.

Also fixed valgrind complaint about uninitialized data in the socket test.